delete user functionality

+++ server/controller/product/deleteProduct.js
@@ -1,7 +1,15 @@
 const productModel = require("../../models/productModel");
+const deleteProductPermission = require("../../helpers/permission");
 
 const deleteProductController = async (req, res) => {
   try {
+    if (!deleteProductPermission(req.userId)) {
+      return res.status(403).json({
+        message: "Permission denied",
+        success: false,
+      });
+    }
+
     const { _id } = req.body;
 
     const deletedProduct = await productModel.findByIdAndDelete(_id);
